+++ server/controllers/api/index.js
@@ -6,8 +6,9 @@ const technicianRoutes = require("./TechnicianRoute");
 const customerRoutes = require("./CustomerRoute");
 const appointmentRoutes = require("./AppointmentRoute");
 const categoryRoutes = require("./CategoryRoute");
-const miscellaneousesRoutes=require("./MiscellaneousRoute");
-const authenticationRoutes=require("./AuthenticationRoute");
+const miscellaneousesRoutes = require("./MiscellaneousRoute");
+const authenticationRoutes = require("./AuthenticationRoute");
+const notificationRoutes = require("./NotificationRoute");
 
 // Use the routes for each model
 router.use("/services", serviceRoutes);
@@ -16,6 +17,7 @@ router.use("/customers", customerRoutes);
 router.use("/appointments", appointmentRoutes);
 router.use("/categories", categoryRoutes);
 router.use("/miscellaneouses", miscellaneousesRoutes);
-router.use("/authentication",authenticationRoutes);
+router.use("/authentication", authenticationRoutes);
+router.use("/notification", notificationRoutes);
 
 module.exports = router;
